Add a command to apply state parts. From disk or S3.
