+++ Cargo.toml
@@ -159,7 +159,7 @@ rand_xorshift = "0.3"
 rayon = "1.5"
 redis = "0.21.5"
 reed-solomon-erasure = "4"
-regex = "1.7.1"
+regex = "1.7.0"
 reqwest = { version = "0.11.0", features = ["blocking"] }
 ripemd = "0.1.1"
 rlimit = "0.7"
