+++ tools/state-viewer/src/cli.rs
@@ -244,7 +244,9 @@ impl ApplyStatePartsCmd {
             home_dir,
             near_config,
             store,
-            state_parts::Location::new(self.root_dir, (self.s3_bucket, self.s3_region)),
+            self.root_dir,
+            self.s3_bucket,
+            self.s3_region,
         );
     }
 }
@@ -427,7 +429,9 @@ impl DumpStatePartsCmd {
             home_dir,
             near_config,
             store,
-            state_parts::Location::new(self.root_dir, (self.s3_bucket, self.s3_region)),
+            self.root_dir,
+            self.s3_bucket,
+            self.s3_region,
         );
     }
 }
