+++ tools/state-viewer/src/lib.rs
@@ -4,11 +4,10 @@ mod apply_chain_range;
 mod apply_chunk;
 pub mod cli;
 mod commands;
-mod contract_accounts;
-mod dump_state_parts;
 mod epoch_info;
 mod rocksdb_stats;
 mod state_dump;
+mod state_parts;
 mod tx_dump;
 
 pub use cli::StateViewerSubCommand;
